refactor(send): extract tc-token lifecycle and pin/revoke actions into submodules
Convert the src/send.rs god-module into a src/send/ directory, moving two
cohesive, self-contained method clusters out of the single large impl Client
block into sibling modules:

- src/send/tctoken_lifecycle.rs: the TC-token lifecycle (issue/cache/fetch/
  persist/re-issue) methods.
- src/send/actions.rs: the pin / unpin / revoke / keep message actions.

The core send path (send_message, send_message_impl, status/reaction helpers,
SKDM cache, phash, SendError/SendResult, tests) stays in src/send/mod.rs.

Pure module split: methods are moved verbatim. The only edits are three
private tc-token methods bumped to pub(super) so the core send path in mod.rs
can still call them across the module boundary. No logic, signature, or
behavior changes.
